+++ serial_interface.h
@@ -1,7 +1,7 @@
 
 #ifndef _SERIAL_INTERFACE_H
 #define _SERIAL_INTERFACE_H
-void getFromSerial(int * serialData);
-void EEPROMLoadConfig(int *temp, int *correct);
-void EEPROMSaveConfig(int temp, int correct);
+void getFromSerial(float * serialData);
+void EEPROMLoadConfig(int *temp, float *correct);
+void EEPROMSaveConfig(int temp, float correct);
 #endif
